+++ extensions/homeassistant/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Home Assistant Changelog
 
+## [Modernize] - 2023-07-16
+
+- Add support for switching assist pipelines
+
 ## [Fix] - 2023-06-25
 
 - Use nearest resolve url also for http requests